GeometryProcessing - Cleaning up implicit type cast warnings in DynamicMesh-based selections, simplification and other processing elements.
#preflight 62bc9eb4a328a968bafc368a
#rb Jimmy.Andrews

[CL 21007309 by David Hill in ue5-main branch]

+++ Engine/Plugins/Runtime/GeometryProcessing/Source/DynamicMesh/Private/MeshSimplification.cpp
@@ -770,7 +770,7 @@ void TMeshSimplification<QuadricErrorType>::SimplifyToMaxError(double MaxError)
 	SimplifyMode = ETargetModes::MaxError;
 	TargetCount = 1;
 	MinEdgeLength = FMathd::MaxReal;
-	MaxErrorAllowed = MaxError;
+	MaxErrorAllowed = (float)MaxError;
 	DoSimplify();
 }
